pyproject.toml
[tool.poetry]
name = "portodjangostarter"
version = "0.1.1"
description = "With portodjangostarter we can start porto containers in django project"
authors = ["BakdauletBolat"]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.10.0"
Django = "^3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
